Added FormCheckpoints

+++ src/main/java/com/sg/fsp/repository/FormRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormRepository extends JpaRepository<Form, Integer> {
+public interface FormRepository extends JpaRepository<Form, Long> {
     Form findFormByFormCode(String formCode);
 }
